[WIP] Refactor events in different files and enable Comment event editing
